Extract copy core factory
