+++ notes/copy-functions.md
@@ -24,3 +24,9 @@
 - `copyPresenterFiles(dirs, io, messageLogger)` – inspects directories, enumerates files, and copies them with filesystem helpers.
 - `copySupportingDirectories(dirs, io, messageLogger)` – orchestrates multiple directory copy plans with filesystem helpers.
 - `main(io, messageLogger)` – orchestrates the workflow, invoking filesystem-dependent helpers and `path` utilities indirectly.
+
+## Next steps for isolating the core
+- Extract the pure helper factory (`createCopyCore`) into its own module so adapter scripts can import it without dragging in Node-specific setup.
+- Keep the existing `directories`, `path`, and `io` objects in the adapter entry point (`copy.js`) and pass them to the shared factory.
+- Once the factory is externalized, progressively move the pure helpers into `src/core` while leaving filesystem orchestration inside `src/generator` until the new adapter folder is ready.
+- Introduce focused unit tests around the exported factory to protect the pure logic during the upcoming directory reshuffle.
